Move message routing to own source files
