+++ mcu/nbproject/configurations.xml
@@ -26,6 +26,8 @@
                    projectFiles="false">
       <itemPath>Makefile</itemPath>
     </logicalFolder>
+    <itemPath>src/routing.c</itemPath>
+    <itemPath>src/routing.h</itemPath>
   </logicalFolder>
   <sourceRootList>
     <Elem>src</Elem>
