+++ mcu/src/routing.h
@@ -1,0 +1,10 @@
+#ifndef ROUTING_H
+#define ROUTING_H
+
+#include "buffer.h"
+
+void check_frame (circular_buf * );
+
+void handle_message(circular_buf * );
+
+#endif
